added vault object
